fix(kubernetes): handle missing $KUBECONFIG
